Smoothening out
Allowed user to enter lower case answers with .upper(), user is now
prompted to enter how many nights the guest will be staying. Eventually
I'd like for the program to show the end date of the guest's stay, or
even have the user enter the start and end dates of a stay.